test: centralize temp_cache fixture in conftest.py (Q11.8)
Four byte-identical copies across test files collapse into one shared
fixture.
