+++ tests/test_cache.py
@@ -11,32 +11,8 @@ import sqlite3
 import tempfile
 from pathlib import Path
 
-import pytest
-
 from brasa.engine.cache import CacheManager, CacheMetadata, _extract_http_status
 from brasa.util import DownloadArgs
-
-
-@pytest.fixture
-def temp_cache():
-    """Create a temporary cache directory for testing."""
-    with tempfile.TemporaryDirectory() as tmpdir:
-        original_cache = CacheManager.__dict__.get("__it__")
-        CacheManager.__it__ = None
-
-        cache = CacheManager()
-        cache._cache_folder = tmpdir
-        Path(tmpdir).mkdir(parents=True, exist_ok=True)
-        Path(cache.cache_path(cache._meta_folder)).mkdir(parents=True, exist_ok=True)
-        Path(cache.cache_path(cache._db_folder)).mkdir(parents=True, exist_ok=True)
-        cache.create_meta_db()
-
-        yield cache
-
-        if original_cache is not None:
-            CacheManager.__it__ = original_cache
-        else:
-            CacheManager.__it__ = None
 
 
 class TestSaveTrialWithStatus:
