+++ tests/test_download_status.py
@@ -4,14 +4,11 @@ Tests cover all 7 core outcomes: . (PASSED), F (FAILED), E (ERROR),
 S (SKIPPED), D (DUPLICATED), I (INVALID), C (CORRUPTED).
 """
 
-import tempfile
 from contextlib import closing
 from pathlib import Path
 
-import pytest
-
 from brasa.engine.api import _should_download
-from brasa.engine.cache import CacheManager, CacheMetadata
+from brasa.engine.cache import CacheMetadata
 from brasa.engine.exceptions import (
     CorruptedContentException,
     DownloadException,
@@ -24,28 +21,6 @@ from brasa.engine.reporting import (
     map_exception_to_download_status,
     to_task_status,
 )
-
-
-@pytest.fixture
-def temp_cache():
-    """Create a temporary cache directory for testing."""
-    with tempfile.TemporaryDirectory() as tmpdir:
-        original_cache = CacheManager.__dict__.get("__it__")
-        CacheManager.__it__ = None
-
-        cache = CacheManager()
-        cache._cache_folder = tmpdir
-        Path(tmpdir).mkdir(parents=True, exist_ok=True)
-        Path(cache.cache_path(cache._meta_folder)).mkdir(parents=True, exist_ok=True)
-        Path(cache.cache_path(cache._db_folder)).mkdir(parents=True, exist_ok=True)
-        cache.create_meta_db()
-
-        yield cache
-
-        if original_cache is not None:
-            CacheManager.__it__ = original_cache
-        else:
-            CacheManager.__it__ = None
 
 
 class TestPassedStatus:
